Egor/speaker_service/requirements.txt
redis>=5.0,<6.0
torch==2.5.1
torchaudio==2.5.1
pyannote.audio>=3.1,<4.0
huggingface_hub>=0.20,<0.24
scipy>=1.11
numpy>=1.24
